feat(logging): refactor logging structure and enhance sensitive data redaction
