+++ server/src/modules/logging/service.ts
@@ -2,15 +2,38 @@ import { Injectable, LoggerService } from '@nestjs/common';
 import { RequestContext } from '@modules/request-context/service';
 import { Logger as PinoBaseLogger } from 'pino';
 import { ignoreLogPaths } from '../logging/constant';
-import { redactSensitiveKeys } from '@helpers/log-redaction.helper';
 import { buildBaseLogger } from './base-logger';
+
+const SENSITIVE_KEYS = [
+  'password',
+  'current_password',
+  'new_password',
+  'token',
+  'invitation_token',
+  'access_token',
+  'refresh_token',
+  'secret',
+  'api_key',
+  'authorization',
+  'cookie',
+];
+
+// Redact known-sensitive keys before an object gets flattened into the log message string —
+// has to happen here, not downstream, since there's no object shape left to redact after.
+function redact(value: unknown, seen = new WeakSet<object>()): unknown {
+  if (typeof value !== 'object' || value === null) return value;
+  if (seen.has(value)) return '[CIRCULAR]';
+  seen.add(value);
+  const result: Record<string, unknown> = {};
+  for (const [key, val] of Object.entries(value)) {
+    result[key] = SENSITIVE_KEYS.includes(key.toLowerCase()) ? '[REDACTED]' : redact(val, seen);
+  }
+  return result;
+}
 
 @Injectable()
 export class TransactionLogger implements LoggerService {
-  // Shared with nestjs-pino's HTTP/bootstrap logger (loader.ts) — buildBaseLogger() is
-  // memoized at module scope, so both consumers resolve to the exact same pino instance
-  // regardless of which one happens to call it first. TransactionLogger's only job here is
-  // route/transactionId enrichment on top of it, not owning construction.
+  // Shared with nestjs-pino's HTTP/bootstrap logger (loader.ts) via buildBaseLogger().
   private readonly logger: PinoBaseLogger = buildBaseLogger();
 
   private enrichLogData(
@@ -25,12 +48,18 @@ export class TransactionLogger implements LoggerService {
     const transactionId = RequestContext.getTransactionId();
     const route = RequestContext.getRoute();
     const startTime = RequestContext.getStartTime();
-    // Redact BEFORE stringify, not after — once an object param is flattened into the
-    // message string, there's no structure left for any path-based redactor (pino's own
-    // `redact` included) to match against. This is the only point in the pipeline where
-    // the original object shape still exists.
+    // A logging call must never crash its caller — if redact/stringify blows up on some
+    // exotic value (a throwing getter, a circular ref past the seen-set, etc.), fall back
+    // to a placeholder for that one param instead of losing the whole log line.
     const formattedParams = optionalParams
-      .map((param) => (typeof param === 'object' ? JSON.stringify(redactSensitiveKeys(param)) : param))
+      .map((param) => {
+        if (typeof param !== 'object' || param === null) return param;
+        try {
+          return JSON.stringify(redact(param));
+        } catch {
+          return '[unloggable]';
+        }
+      })
       .join(' ')
       .trim();
